move merge value_traits.hpp into impl/conversion.hpp
